Implement self vote part of candidate role.
Next step is to implement the request vote part, for which I'll need
to stub out the `send` function.

The treatment of role-specific state I'm using here is a little
clumsy. An improvement could be to e.g. create
`modifyVolatileCandidateState` that checks that the role is
`Candidate` and that the volatile candidate state is a `Just` and
fails otherwise. This allows most of the code that manipulates
role specific states to assume the role specific state exists, and not
have to deal with the `Maybe`.