Implement startup code.
I'm now able to run servers! They don't do much yet, since so far I
don't actually let them talk to each other (but everything I need for
that is set up).

I put a bit of effort into setting up optional debug messages via a
central queue, and explicitly killing all threads on shutdown. By
default, the GHC RTS kills all remaining threads when the "main"
thread exits, but this doesn't seem to happen when you run `:main` in
GHCi and then kill it with `CTRL-C`. Since I test by running `:main`
in GHCi, the explicit killing is useful, but it clutters up the code a
little.

To convince myself the thread killing was working, I added some
messages in the `finally` clauses, but these were appearing on the
console garbled and overlapped, since they all happen at about the
same time. So, I implemented a thread safe `printf` using a lock. This
obviates the need for queue for debug messages, so I should probably
go back and simplify that code a little by removing the queue ...

I also set the min election timeout to a very large value, because
it's too hard to tell what's going on when using a small value as
suggested in the paper.